+++ hutool-http/src/main/java/cn/hutool/http/GlobalInterceptor.java
@@ -1,7 +1,8 @@
 package cn.hutool.http;
 
 /**
- * 全局的拦截器
+ * 全局的拦截器<br>
+ * 包括请求拦截器和响应拦截器
  *
  * @author looly
  * @since 5.8.0
@@ -9,34 +10,81 @@ package cn.hutool.http;
 public enum GlobalInterceptor {
 	INSTANCE;
 
-	private final HttpInterceptor.Chain interceptors = new HttpInterceptor.Chain();
+	private final HttpInterceptor.Chain<HttpRequest> requestInterceptors = new HttpInterceptor.Chain<>();
+	private final HttpInterceptor.Chain<HttpResponse> responseInterceptors = new HttpInterceptor.Chain<>();
 
 	/**
 	 * 设置拦截器，用于在请求前重新编辑请求
 	 *
 	 * @param interceptor 拦截器实现
 	 */
-	synchronized public GlobalInterceptor addInterceptor(HttpInterceptor interceptor) {
-		this.interceptors.addChain(interceptor);
+	synchronized public GlobalInterceptor addRequestInterceptor(HttpInterceptor<HttpRequest> interceptor) {
+		this.requestInterceptors.addChain(interceptor);
 		return this;
 	}
 
 	/**
-	 * 清空
+	 * 设置拦截器，用于在响应读取后完成编辑或读取
+	 *
+	 * @param interceptor 拦截器实现
+	 */
+	synchronized public GlobalInterceptor addResponseInterceptor(HttpInterceptor<HttpResponse> interceptor) {
+		this.responseInterceptors.addChain(interceptor);
+		return this;
+	}
+
+	/**
+	 * 清空请求和响应拦截器
+	 *
 	 * @return this
 	 */
-	synchronized public GlobalInterceptor clear(){
-		interceptors.clear();
+	public GlobalInterceptor clear() {
+		clearRequest();
+		clearResponse();
 		return this;
 	}
 
 	/**
-	 * 复制过滤器列表
+	 * 清空请求拦截器
+	 *
+	 * @return this
+	 */
+	synchronized public GlobalInterceptor clearRequest() {
+		requestInterceptors.clear();
+		return this;
+	}
+
+	/**
+	 * 清空响应拦截器
+	 *
+	 * @return this
+	 */
+	synchronized public GlobalInterceptor clearResponse() {
+		responseInterceptors.clear();
+		return this;
+	}
+
+	/**
+	 * 复制请求过滤器列表
+	 *
 	 * @return {@link cn.hutool.http.HttpInterceptor.Chain}
 	 */
-	HttpInterceptor.Chain getCopied(){
-		final HttpInterceptor.Chain copied = new HttpInterceptor.Chain();
-		for (HttpInterceptor interceptor : this.interceptors) {
+	HttpInterceptor.Chain<HttpRequest> getCopiedRequestInterceptor() {
+		final HttpInterceptor.Chain<HttpRequest> copied = new HttpInterceptor.Chain<>();
+		for (HttpInterceptor<HttpRequest> interceptor : this.requestInterceptors) {
+			copied.addChain(interceptor);
+		}
+		return copied;
+	}
+
+	/**
+	 * 复制响应过滤器列表
+	 *
+	 * @return {@link cn.hutool.http.HttpInterceptor.Chain}
+	 */
+	HttpInterceptor.Chain<HttpResponse> getCopiedResponseInterceptor() {
+		final HttpInterceptor.Chain<HttpResponse> copied = new HttpInterceptor.Chain<>();
+		for (HttpInterceptor<HttpResponse> interceptor : this.responseInterceptors) {
 			copied.addChain(interceptor);
 		}
 		return copied;