+++ hutool-http/src/main/java/cn/hutool/http/HttpResponse.java
@@ -34,6 +34,10 @@ import java.util.Map.Entry;
 public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 
 	/**
+	 * Http配置
+	 */
+	protected HttpConfig config;
+	/**
 	 * 持有连接对象
 	 */
 	protected HttpConnection httpConnection;
@@ -62,13 +66,15 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 	 * 构造
 	 *
 	 * @param httpConnection {@link HttpConnection}
+	 * @param config         Http配置
 	 * @param charset        编码，从请求编码中获取默认编码
 	 * @param isAsync        是否异步
 	 * @param isIgnoreBody   是否忽略读取响应体
 	 * @since 3.1.2
 	 */
-	protected HttpResponse(HttpConnection httpConnection, Charset charset, boolean isAsync, boolean isIgnoreBody) {
+	protected HttpResponse(HttpConnection httpConnection, HttpConfig config, Charset charset, boolean isAsync, boolean isIgnoreBody) {
 		this.httpConnection = httpConnection;
+		this.config = config;
 		this.charset = charset;
 		this.isAsync = isAsync;
 		this.ignoreBody = isIgnoreBody;
@@ -273,7 +279,7 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 		Assert.notNull(out, "[out] must be not null!");
 		final long contentLength = contentLength();
 		try {
-			return copyBody(bodyStream(), out, contentLength, streamProgress);
+			return copyBody(bodyStream(), out, contentLength, streamProgress, this.config.ignoreEOFError);
 		} finally {
 			IoUtil.close(this);
 			if (isCloseOut) {
@@ -563,7 +569,7 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 
 		final long contentLength = contentLength();
 		final FastByteArrayOutputStream out = new FastByteArrayOutputStream((int) contentLength);
-		copyBody(in, out, contentLength, null);
+		copyBody(in, out, contentLength, null, this.config.ignoreEOFError);
 		this.bodyBytes = out.toByteArray();
 	}
 
@@ -572,13 +578,14 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 	 * 异步模式下直接读取Http流写出，同步模式下将存储在内存中的响应内容写出<br>
 	 * 写出后会关闭Http流（异步模式）
 	 *
-	 * @param in             输入流
-	 * @param out            写出的流
-	 * @param contentLength  总长度，-1表示未知
-	 * @param streamProgress 进度显示接口，通过实现此接口显示下载进度
+	 * @param in               输入流
+	 * @param out              写出的流
+	 * @param contentLength    总长度，-1表示未知
+	 * @param streamProgress   进度显示接口，通过实现此接口显示下载进度
+	 * @param isIgnoreEOFError 是否忽略响应读取时可能的EOF异常
 	 * @return 拷贝长度
 	 */
-	private static long copyBody(InputStream in, OutputStream out, long contentLength, StreamProgress streamProgress) {
+	private static long copyBody(InputStream in, OutputStream out, long contentLength, StreamProgress streamProgress, boolean isIgnoreEOFError) {
 		if (null == out) {
 			throw new NullPointerException("[out] is null!");
 		}
@@ -588,7 +595,7 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 			copyLength = IoUtil.copy(in, out, IoUtil.DEFAULT_BUFFER_SIZE, contentLength, streamProgress);
 		} catch (IORuntimeException e) {
 			//noinspection StatementWithEmptyBody
-			if (HttpGlobalConfig.isIgnoreEOFError()
+			if (isIgnoreEOFError
 					&& (e.getCause() instanceof EOFException || StrUtil.containsIgnoreCase(e.getMessage(), "Premature EOF"))) {
 				// 忽略读取HTTP流中的EOF错误
 			} else {