+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### ❌不兼容特性
 * 【extra  】     【可能兼容问题】BeanCopierCache的key结构变更
+* 【http   】     【可能兼容问题】HttpInterceptor增加泛型标识，HttpRequest中配置汇总于HttpConfig
 
 ### 🐣新特性
 * 【core   】     MapUtil增加entry、ofEntries方法
@@ -13,6 +14,7 @@
 * 【core   】     IterUtil增加filtered，增加FilterIter（issue#2228）
 * 【core   】     增加NodeListIter、ResettableIter
 * 【crypto 】     HmacAlgorithm增加SM4CMAC（issue#2206@Github）
+* 【http   】     增加HttpConfig，响应支持拦截（issue#2217@Github）
 
 ### 🐞Bug修复
 * 【core   】     IdcardUtil#getCityCodeByIdCard位数问题（issue#2224@Github）